requirements.txt
absl-py==0.9.0
astor==0.8.1
atomicwrites==1.3.0
attrs==19.3.0
boto3==1.9.238
botocore==1.12.238
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
docutils==0.15.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.8.4
grpcio==1.27.1
h5py==2.10.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.5.0
jmespath==0.9.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
libsixel-python==0.5.0
Markdown==3.2
more-itertools==8.2.0
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.32
opt-einsum==3.1.0
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
pmdarima==1.3.0
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pydantic==0.28
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.1
PySide2==5.14.1
pytest==5.3.5
python-dateutil==2.8.1
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
scipy==1.4.1
shiboken2==5.14.1
six==1.14.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
tensorflow-hub==0.7.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.36.1
ujson==1.35
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.11.2
zipp==2.2.0
